refactor: multiple - move add_member to admin.py - fix signup link - show account as email while signing up
