+++ frontend/src/pages/dashboard/Members.vue
@@ -128,7 +128,6 @@ const dropdownOptions = (name, isAdmin) => {
 const removeMemberOptions = computed(() => ({
 	title: __('Remove Member'),
 	message: __(`Are you sure you want to remove member ${memberToBeRemoved.value}?`),
-	size: 'xl',
 	icon: {
 		name: 'alert-triangle',
 		appearance: 'warning',
